requirements.txt
transformers == 2.1.1
pytorch == 1.3.0